fixup stuck in systemctl stop syswatcher

+++ core/cleanup.c
@@ -4,7 +4,10 @@
 #include <stdio.h>
 #include <plugin_server.h>
 #include <log.h>
+#include <pthread.h>
 
+pthread_mutex_t clean_mutex = PTHREAD_MUTEX_INITIALIZER;
+int clean_done = 0;
 void delete_metric(struct metric_unit *unit) {
     unit->do_del_metric(unit);
 }
@@ -25,15 +28,20 @@ void stop_collector(void) {
 
 void cleanup(void)
 {
-    logging(LEVEL_ZERO, "AT EXIT\n");
-    logging(LEVEL_ZERO, "stop collector\n");
-    stop_collector();
-    logging(LEVEL_ZERO, "delete all metric\n");
-    delete_all_metric();
-    logging(LEVEL_ZERO, "unload plugin server\n");
-    plugin_server_finish();
-    logging(LEVEL_ZERO, "clean syswatcher\n");
-    exit_syswatcher(&watcher);
-    logging(LEVEL_ZERO, "DONE\n");
-    exit_logger(&log_unit);
+    pthread_mutex_lock(&clean_mutex);
+    if (!clean_done) {
+        logging(LEVEL_ZERO, "AT EXIT\n");
+        logging(LEVEL_ZERO, "stop collector\n");
+        stop_collector();
+        logging(LEVEL_ZERO, "delete all metric\n");
+        delete_all_metric();
+        logging(LEVEL_ZERO, "unload plugin server\n");
+        plugin_server_finish();
+        logging(LEVEL_ZERO, "clean syswatcher\n");
+        exit_syswatcher(&watcher);
+        logging(LEVEL_ZERO, "DONE\n");
+        exit_logger(&log_unit);
+        clean_done = 1;
+    }
+    pthread_mutex_unlock(&clean_mutex);
 }
